pyproject.toml
[tool.poetry]
name = "streams"
version = "0.1.0"
description = ""
authors = ["Professional <pyostr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
redis = "^6.4.0"
python-dotenv = "^1.1.1"
pybit = "^5.12.0"
protobuf = "^6.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
